YH-1330: move onboarding modal to main layout
