+++ src/app/layouts/MainLayout/ui/MainLayout.tsx
@@ -3,6 +3,7 @@ import { Navigate, Outlet, useLocation } from 'react-router-dom';
 
 import { ROUTES } from '@/shared/config/router/routes';
 import { SentryRouteTracker } from '@/shared/config/sentry/SentryRouteTracker';
+import { useAppSelector, useModal } from '@/shared/hooks';
 import { AutoScrollToTop } from '@/shared/ui/AutoScrollToTop';
 import { Breadcrumbs } from '@/shared/ui/Breadcrumbs';
 import { Drawer } from '@/shared/ui/Drawer';
@@ -10,8 +11,10 @@ import { ErrorBoundary } from '@/shared/ui/ErrorBoundary';
 import { ErrorElement } from '@/shared/ui/ErrorElement';
 
 import { listAdminRoles, useProfileQuery } from '@/entities/auth';
+import { getIsEmptySpecialization, getProfilesLength } from '@/entities/profile';
 
 import { Header } from '@/widgets/Header';
+import { OnboardingModal } from '@/widgets/Main/OnboardingModal';
 import { MenuItem, Sidebar } from '@/widgets/Sidebar';
 
 import SkeletonGenerator from '../model/helper/SkeletonGenerator';
@@ -26,9 +29,15 @@ interface MainLayoutProps {
 
 export const MainLayout = ({ sidebarItems, onlyAdmin }: MainLayoutProps) => {
 	const [isOpenSidebarDrawer, setIsOpenSidebarDrawer] = useState(false);
+	const { isOpen, onOpen, onClose } = useModal();
+
 	const location = useLocation();
 
 	const { data: profile, isLoading } = useProfileQuery();
+
+	const profilesCount = useAppSelector(getProfilesLength);
+	const isSpecializationEmpty = useAppSelector(getIsEmptySpecialization);
+
 	const isAdminRole = profile?.userRoles?.some((role) =>
 		listAdminRoles.find((i) => i == role.name),
 	);
@@ -46,6 +55,12 @@ export const MainLayout = ({ sidebarItems, onlyAdmin }: MainLayoutProps) => {
 			setIsOpenSidebarDrawer(false);
 		}
 	}, [location]);
+
+	useEffect(() => {
+		if (isSpecializationEmpty && profilesCount === 1) {
+			onOpen();
+		}
+	}, [isSpecializationEmpty, onOpen, profilesCount]);
 
 	if (isLoading) return <MainLayoutSkeleton />;
 
@@ -97,6 +112,7 @@ export const MainLayout = ({ sidebarItems, onlyAdmin }: MainLayoutProps) => {
 					</Drawer>
 				)}
 			</AutoScrollToTop>
+			<OnboardingModal isOpen={isOpen} onClose={onClose} />
 		</Suspense>
 	);
 };
